vscode: added integrated support for VSCODE
Added support for MS VSCODE in a form of a workspace file
(snapd.code-workspace) configred to develop `snapd` using this IDE,
including recommendations for the set of usefull extensions.
The IDE will suggest installing these extensions when the workspace
file is opened first time.
The IDE gets configured also to provide a simple way of debugging
and running the tests directly via the IDE.

Signed-off-by: Arseniy Aharonov <arseniy.aharonov@canonical.com>