cmd/snap-confine: mount bpffs under /sys/fs/bpf is needed
Some systems out there are a weird mix of older userland and new kernels. We
would normally expect /sys/fs/bpf to be a mount point of bpffs but, it is
possible that the system is configured in such a way that it uses cgroupv2 by
default, but the mount is not present. In this case, we try to mount bpffs
ourselves, and just die if that fails.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>