gadget, gadgettest: reimplement tests to use new gadgettest examples.go file
Add an examples.go file with basic examples of the structures we expect to see
and use in many gadget tests that have to do with MockDiskMapping, etc.

Also move some things around so that Err* variables from device_linux.go are
also available on darwin too, since we need to start importing bits of this
package from the snap command soon as well.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>