i/cups: prevent using the host cupsd sock by not using the AA cups abstractions
This is because by design we want snaps using the cups interface to _only_ have
access to the snapped version of cupsd's socket at /var/cups/cups.sock since we
know that is always safe for snap apps to talk to, as it will always implement
mediation.

We do however grant access to ~/.cups/lpoptions since this is where things like
default printer, etc. are configured and we do want client snaps to continue to
be able to read this.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>