exportmanager: add export manifest
This commit adds the first part of the new export manager. It just
adds the `Manifest` datatype and the handling of exporting and
unexporting files and the special handling for the snapd tooling.

The goal is that eventually there will be a full manager that
is responsible for exporting parts of snaps into a new
```
/var/lib/snapd/export/<snapName>/<exportedVersion>/<exportSet>
```
location. The main use case for this is to export the snapd
tooling to a fixed location
```
/var/lib/snapd/export/snapd/current/tools
```
and snap-confine will the use these tools from the new location.

This is split from https://github.com/snapcore/snapd/pull/9546
and slightly simplified.

Ref: https://forum.snapcraft.io/t/12139