Allow numbers in hook (except for the first character) when checking hook security tag in snap-confine to allow interface hooks for such plugs/slots; this matches the validHook regex (snap/naming/validate.go) used by snapd. Hooks for plugs/slots names that end with numbers otherwise cause an error if snapctl is used:
connect-plug-station0": error: error running snapctl: interface attributes can only be read during the execution of interface hooks

The error arises because snap-confine doesn't consider passed security context as hook's, and overrides SNAP_COOKIE env var
with the one for the snap, replacing the cookie value passed by snap run
for the hook execution. As a consequence snapctl considers to be running
with an ephemeral context (not from the hook), so cannot figure hook
name etc and errors out.