gadget/install/partition.go: include DiskIndex in synthesized OnDiskStructure
This is not strictly necessary, but resolves an issue wherein we previously
were mounting all partitions during install mode at the same location of
<contentMountpoint>/0 in writeFilesystemContent(). This is because we were 
using ds.Index as the directory to use for each partition, but now that we
switched to using YamlIndex and DiskIndex, the mount code in 
writeFilesystemContent() was adapted to use DiskIndex which made sense at the 
time because the structure it is operating on is an OnDiskStructure. This 
didn't cause any obvious issues since we unmount things before mounting again
so even though we would be mounting on the same directory different partitions, 
nothing was broken.

The alternative to this change is to revert to effectively using the YamlIndex
instead as the final mount point directory name, but again it makes more sense
to use the DiskIndex in writeFilesystemContent() because it is accepting an
OnDiskStructure instead of a LaidOutStructure.

The other reason we didn't notice this is because we do not have any end to end
tests of install.Run() where we would have noticed that 
createMissingPartitions() was providing incorrectly synthesized 
OnDiskStructures to be used with writeFilesystemContent(). That should be 
remedied soon (and indeed I only noticed this while writing unit tests for 
install.Run()).

This bug was introduced in snapcore/snapd#11234.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>