cmd/snap/auto-import: use osutil.LoadMountInfo impl instead
Eliminate the custom mount table parsing here in lieu of the existing and
arguably better tested parsing in LoadMountInfo.

Also fix a test-only bug where we were not properly skipping the UC20+ specific
mount dir of /writable/system-data/var/lib/snapd/seed. The code as written only
worked when dirs.GlobalRootDir was the empty string, if that was any other dir
like in tests, then it wouldn't match and we would consider it. However the
unit test for this specific case, TestAutoImportUC20CandidatesIgnoresSystemPartitions
had inadvertently copied the erroneous mount line which did not have a "-" to
delineate between the optional flags for the mount entry and the rest of the
mount fields which was used in another test, so it was silently skipped with
the custom implementation that was here before. Fixing this line demonstrated
the test-only issue mentioned above and so it had to be fixed to properly test
this case.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>