devicestate: wait for NTP sync before trying to get a serial assertion
One of the recent bugreport we got was that on systems with no
battery powered clocks the device registration can be slow. The
first step here is to ensure that any logic we run that involves
backoff/timers waits for NTP synchronization. This commit adds
some simple code to ensure this.