o/servicestate: introduce EnsureQuotaUsability and use it
This needs to be exported for other package unit tests which use / interact
with quotas, since otherwise the systemd error (if non-nil such as when running
the unit tests on trusty and xenial) will be set too early before a newer
systemd version can be mocked in a given unit test. Having this function return
a restoring function ensures that calling this function to re-calculate the
errors (with the presumably mocked values) is undo-able for other unit tests.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>