tests: add test for running a snap command from a devmode snap
This commit adds a test that checks that we can run a snap command
from inside a devmode snap.

Because of historic mistakes we allowed this and until we properly
deprecated it we need to ensure it works. We really do not want to
support running other snaps from devmode snaps as the use-case for
devmode is to get help on the way to confined snaps. But snaps can
not run other snaps so this does not make sense.

Signed-off-by: Michael Vogt <michael.vogt@canonical.com>

tests/main/snap-run-devmode: rename, adjust to only test classic

Revamp the main suite test to only test the classic use cases for running snaps
from devmode snaps, and follow the pattern of
tests/nested/manual/devmode-snaps-can-run-other-snaps.

Running on xenial classic affords us the ability to avoid messing with lxd and
get some speedup there.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>