data/env: fix fish env for all versions of fish, unexport local vars, export XDG_DATA_DIRS
Make the fish env setup script compatible with old versions of fish 3.0.x which
was released back in 2019, but is still shipped by recent releases of distros
eg. Ubuntu 21.10.

While doing so, make sure that helper variables used in the script stay local.

Since only newer versions of fish (3.1+) are aware of XDG_DATA_DIRS make sure
that it's exported.

Fixes:
https://bugs.launchpad.net/ubuntu/+source/snapd/+bug/1958022
https://bugs.launchpad.net/ubuntu/+source/snapd/+bug/1957948
https://bugs.launchpad.net/ubuntu/+source/snapd/+bug/1960492

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>