gadget: make encryption options a map, rename and adjust validation logic
We now expect a map of potential options, with the only currently used one as
"method", which is required in the map. This lets us better future proof if we
need to add other keys to this map in the future.

This also means we can get rid of the type alias DiskEncryptionMethod as we now
just use strings for anything anyways.

Also adjust some doc-comments.

Thanks to Alberto and Samuele for these suggestions.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>