tests/main/uc20-create-partitions: add debugging information
It occurred to me that what might be happening when we see this test fail is
not that networking has disappeared and thus we lose the connection to the
spread machine, but rather than we are stuck/hung/deadlocked in
uc20-create-partitions.

So add some debug output there if we get stuck for 3 minutes. This should be
plenty long enough to actually run the test, but less than the time it takes
for the spread connection to time out.

For reference, the uc20-create-partitions-encrypt variant of the test, which
should take the longest to execute, takes less than 2 minutes to execute:

2022-02-11 22:25:25 Preparing google:ubuntu-20.04-64:tests/main/uc20-create-partitions-encrypt (feb112148-627987)...
...
2022-02-11 22:25:33 Executing google:ubuntu-20.04-64:tests/main/uc20-create-partitions-encrypt (feb112148-627987) (274/566)...
...
2022-02-11 22:26:21 Restoring google:ubuntu-20.04-64:tests/main/uc20-create-partitions-encrypt (feb112148-627987)...


Signed-off-by: Ian Johnson <ian.johnson@canonical.com>