packaging/ubuntu-16.04/rules: don't run unit tests on riscv64
The machines running the tests are not currently fast enough to reliably pass
or be trusted, so for now just disable the tests. We will re-enable them when
we have time to adjust the tests for real or when we have better hardware that
can actually complete tests in a reasonable amount of time.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>