gadget: Use systemd-repart to create new partition during install
Instead of laying out partition and creating manually, we convert the
`gadget.yaml` volume structure into a `repart.d` configuration.

This is experimental, and we fallback to the old code if an error
occurs.

This requires `systemd-repart` to be available in the core snap.
To build such a core snap, a rebuild of the systemd Ubuntu package
is required enabling `-Drepart=true`.