gadget: add buildVolumeStructureToLocation, volumeStructureToLocationMap
These functions will replace much of the device_linux.go functions we currently
use to identify where to find a structure to update it.

Also introduce a new union type, StructureLocation which is used to identify
where to write updates for an asset, depending on whether it has a filesystem
or not.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>