cmd/snap-bootstrap: support booting into factory-reset mode
Add support for setting up the system in factory-reset mode. We only care about
ubuntu-save, as ubuntu-data and ubuntu-boot will be wiped. As such, it is
expected that only the fallback sealed key object has been sealed for this mode.
However, we also allow to prompt the user for the recovery (aka reinstall) key.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>