run-checks: replace linters w/ golangci
Since we're now using golangci in Github, this adds it to run-checks
and removes the calls to linters which are covered by golangci.
Unfortunately, we can't call it w/ --disable due the current config
so the output is filtered to ignore certain linters, if the SKIP_*
env vars are used. Also unfortunately, a 'go get' of golangci gets
OOM killed (in Go 1.13 at least), so this gets the binary from Github.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>