cmd/snap-device-helper: make sysroot static
Make sysroot static, since it does not need to be exported anywhere. This was
done initially such that sysroot can be overridden from the tests, but the test
had to include the whole snapd-device-helper.c anyway, so the variable can be
marked as static now.

Sparse warning:

snap-device-helper/snap-device-helper-test.c: note: in included file:
snap-device-helper/snap-device-helper.c:147:12: warning: symbol 'sysroot' was not declared. Should it be static?

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>