gadget/update.go: refactor fatal error handling
Use a better error helper which can return the error directly that eliminates
the if check in the function.

Also add a check for laidOutVol to be non-nil, failing if it is.

Finally, adjust the declaration of the StructureLocation to be outside of the
if and just set it in the map outside of the if.

Thanks to Alberto and Samuele for these suggestions.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>