usersession/userd: ask the user whether to open the URL by given app
PR https://github.com/snapcore/snapd/pull/11200 added a way to query for a
handler for any given URL scheme. Extend this and ask the user whether a URL
should be opened by a given handler. This is similar to what the desktop portal
does, but instead of presenting the app chooser dialog, we only present the user
with a yes/no question and the default handler.

Note, that the implementation is only for the fallback io.snapcraft.Launcher
handler, which would be called only if the desktop portal is not available.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>