tests/main/snapd-snap: test installing the snapd snap and connecting interfaces
This is originally from the work to vendor apparmor 3 into the snapd snap, but
is more useful generally as it ensures we can install/use the snapd snap on
classic.

Having this test here will make it easier to propose some changes to the snapd
snap and ensure those are adequately tested.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>