interface/opengl: allow read on /proc/sys/dev/i915/perf_stream_paranoid
/proc/sys/dev/i915/perf_stream_paranoid is used to allow toggling Intel
Graphics metrics collection to non-root users. The firefox snap (at
least) reads this during startup (but doesn't try to write to it) which
causes log noise but otherwise seems to work ok. Since the contents are
a simple integer value, there is nothing sensitive in the file, so allow
snaps that plug opengl to read the file.

References:
- https://github.com/intel/pti-gpu#prerequisites