tests: switch to "test-snapd-swtpm"
We currently use `swtpm-mvo` for testing. However this snap is
owned by an individual (me) and it's not clear from the name
that the main purpose of the snap is to test snapd.

So this commit moves to the new `test-snapd-swtpm` snap that
is team maintained in `github.com/snapcore/test-snapd-swtpm`
and has a more suitable name.

Thanks to Ian for the rename of the snap and for modernizing it.