Fix the nested remodel tests
Currently the test is failing in core 18 because the core18 snap is
being autore-freshed breaking the connection with the host machine.

The idea of the change is to use the tests.nested tool which is widely
used on nested tests and repacks the core18 snap avoiding
auto-refreshes.

This change also includes a new wait-for event used to wait until the
device is initialized and has a serial. By doing that it is not needed
anymore to source the nested.sh lib.