release-tools/do-release.sh: add prototype script to automate release process
This is a prototype script I use for getting snippets of what to execute to do
things in the release. It is not complete, but there are TODOs left for how to
make it complete if we wanted to try and automate the release process as much
as possible that folks could implement.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>