systemd: remove the systemctl stop timeout handling
This PR is somewhat related to:
https://github.com/snapcore/snapd/pull/11340

In the process of improving the systemd stop code to better track
the progress of stopping units, we also removed the snapd concept
of a timeout period in order to exclusively rely on
'systemctl stop' to return an error after its internal stop
escalation process failed.

We therefore no longer can produce a 'Timeout' error.

This patch removes code in the code base that used to deal
with this 'Timeout' condition.

The idea behind this patch is that we trust 'systemctl stop'
to tell us if it failed to stop a unit within its configured
timeout, and if it fails, there is nothing more we can do but
to return an error.

Signed-off-by: Fred Lotter <fred.lotter@canonical.com>