tests: fix the nested remodel tests (#11539)
* Fix the nested remodel tests

Currently the test is failing in core 18 because the core18 snap is
being autore-freshed breaking the connection with the host machine.

The idea of the change is to use the tests.nested tool which is widely
used on nested tests and repacks the core18 snap avoiding
auto-refreshes.

This change also includes a new wait-for event used to wait until the
device is initialized and has a serial. By doing that it is not needed
anymore to source the nested.sh lib.

* Change call to nested_get_boot_id by tests.nested boot-id

* Same change on boot id for test remodel-cross-store

* Update to use the use the gadget snap from the store

* Minor fixes to use the env vars to download the gadget and kernel snaps

* Allow skip repack for kernel, base and gadget

This is needed to make remodel tests work. This is also neede for other
tests to be able to use kernel and gadget from the store

* Improve cleanup of tests

* Removing snaps installed in other tests