many: change transactional flag to an option
Change the --transactional flag to an option called --transaction with
two possible values, "per-snap" and "all-snaps". The former is the
default, and makes updates/installs transactional per snap, while the
latter makes the API call transactional on the specified set of snaps,
so if one fails all the action is reverted.