overlord: add missing grub.cfg assets mocks in manager_tests.go
The current code will mock the `amd64` architecutre when running
the manager tests but it does not mock the `grub.cfg:static-cmdline`
or the `grub-recovery.cfg:static-cmdline`. This is needed because
the `bootloader/asserts` package registers the cmdline via `init()`
so this happens before the architecture is mocked in the
`manager_test.go` code. Hence the manager tests will fail when
inspecting the commandline on non amd64/i386 machines.

This commit makes the mocking explicit. An alternative approach
would be to move the snippet registration out of `init()` and
run it on first boot but that has it's own problems so I went
with this approach.