interfaces/modem-manager: add support for Cinterion modules
The modem-manager interface includes udev rules for various modems,
including Cinterion devices. Outside of snapd these definitions are
maintained within the ModemManager repository and have been
significantly updated.

This change syncs the udev rules for the modem manager interface with
the upstream for the Cinterion modems. This data is taken directly from
the Modem Manager repository, furthermore it includes rules for PLS63
modem.

As a consequence ModemManager can read all of the port type hints it
needs which is crucial for it to know how to talk to a particular modem.