Remove the file that configures nm as default
In UC18 the /etc/netplan/00-default-nm-renderer.yaml is created when
network manager is installed.

google:ubuntu-core-18-64 .../tests/main/interfaces-network-manager# cat
/etc/netplan/00-default-nm-renderer.yaml
network:
  renderer: NetworkManager

Then, in a following test, if the machine is rebooted, the network is
discofigured and spread cannot connect anymore to the machine.

I could see this error when manually tryed to setup console-conf

Network configuration failed:
            "/etc/netplan/00-snapd-config.yaml:5:7: Error in network
        definition: eth0: 'set-name:' requires 'match:' properties\n
                            addresses: []\n      ^\n"