interfaces/cpu-control: fix apparmor rules of paths with CPU ID
With the original rules, the paths in "/sys/devices/system/cpu/cpuX" are
not writable. AppArmor shows error messages, for instance:
> audit: type=1400 audit(1649300989.268:103): apparmor="DENIED" operation="open" profile="snap.my-snap.init" name="/sys/devices/system/cpu/cpu4/core_ctl/min_cpus" pid=4702 comm="bash" requested_mask="wc" denied_mask="wc" fsuid=0 ouid=0

According to the AppArmor Core Policy Reference[1], the globbing syntax
is not correct.
> * - match zero or more characters at the directory level. When looking at the path as a string this will match every character except /

So, the following rule mean the CPU ID should be from cpu00* to cpu99*
> /sys/devices/system/cpu/cpu[0-9][0-9]*/core_ctl/min_cpus rw

This patch change the path to be from cpu0* to cpu9*.

[1] https://gitlab.com/apparmor/apparmor/-/wikis/AppArmor_Core_Policy_Reference#apparmor-globbing-syntax

Signed-off-by: Robert Liu <robert.liu@canonical.com>