tests: add new "core22-migrate-dirs" test
This is a testcase for a bug that Alfonso found. If one installs
a snap and then refreshes that snap to core22 the hidden dirs
migration will happen. But when the snap is then purged and
the same steps are done again some dirs in the users home
already exit and that makes the current code fail:
```
And install both again
+ snap install --dangerous /home/gopath/src/github.com/snapcore/snapd/tests/lib/snaps/test-snapd-tools/test-snapd-tools_1.0_all.snap
test-snapd-tools 1.0 installed
+ test-snapd-tools.echo foo
foo
+ snap install --dangerous test-snapd-tools_22.0_all.snap
error: cannot perform the following tasks:
- Copy snap "test-snapd-tools" data (cannot move "/root/snap/test-snapd-tools" to "/root/.snap/data/test-snapd-tools": rename /root/snap/test-snapd-tools /root/.snap/data/test-snapd-tools: file exists)
```

This commit contains the a spread test for this.