snap-confine: remove useless call to chown
This chown was added to assign the hostfs directory to the root
user/group back when this part of the code was run as the root user.
Nowadays it's run as an ordinary user, and therefore calling chown here
would fail; luckily for us, the hostfs directory is created as part of
the packaging, so it's always owned by root.

Once we are sure that this conditions are never hit, we might even want
to remove the checks themselves.