cmd/snap-bootstrap: Listen to keyboard added after start and handle switch root
Keyboard drivers and the drivers for the bus they are plugged into are
not necessarily initialized by the time udev has settled cold plug
events. That means we have to read events for new input devices and
read those keyboards.

Also because the daemon is started in initramfs and needs to write
into the file system, it needs to chroot when switch root happens.