secboot: partial reprovision
Add support for partial reprovisioning which happens during factory reset. We
assume that the TPM has been previously provisioned and we have the lockout
authorization key. Thus we provision it again and authorize this operation using
the existing key. Then, when sealing the keys we will first release the
resources associated with PCR handles for each of the keys.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>