interfaces/builtin: remove the name=org.freedesktop.DBus restriction in cups-control AppArmor rules
This seems to have been included erroneously: the org.freedesktop.DBus
bus name is owned by the bus itself rather than any particular peer
connected to the bus. While it makes sense to use in a rule dealing with
communication with the bus, it seems to block legitimate traffic in this
case.