tests/main/nfs-support: be robust against umount failures
The umount command is often failing because the "test" user's
`dbus-daemon` process is still keeping the /home/test directory busy.

systemd-logind will automatically close the session after the test user
has logged out (which happens at the end of the `su` command), but it
allows for a 10 seconds grace period, which we cannot afford to waste.

Therefore, in addition to adding a retry loop to the umount operation,
we also explicitly kill all of the "test" user's processes.