squashfs: improve error reporting when `unsquashfs` fails
The current code in Snap.Unpack() will report a very generic:
` exit status 1` error if `unsquashfs` fails. This commit changes
the code to first look at any unsquashfs errors reported on stderr
and only if none of these are found it will report a (slightly less)
generic error.

This should help diagnose a bug in the pi-kernel refresh where
the Snap.Unpack() of the kernel assets fails.