squashfs: tweak error handling of Snap.Unpack()
Use a multi-writer when detecting errors from unsquashfs and
capture stderr in addition to the newUnsquashfsStderrWriter()
we currently use. This ensures that on modern unsquashfs that
properly returns non-zero exit codes we show a useful error
message. Size of the buffer should not be a concern because
unsquashfs stops early nowdays (unlike the old versions that
would keep iterating over the squashfs and generating a lot
of output).

The behavior of unsquashfs was checked with:
```
mkdir /tmp/tiny
sudo mount -o size=1M -t tmpfs none /tmp/tiny
sudo unsquashfs -f -d /tmp/tiny /var/lib/snapd/snaps/core20_1524.snap
```

Thanks to Maciej for the suggestion with the multi-writer.