gadget: check also mbr type when testing for implicit data partition
When testing if a gadget defines implicitly the system data
partition (which can happen for UC16 and UC18), we were checking the
type of the on-disk partition to determine if it was the system data
partition. However, we were testing only if it matched the GPT UUID
partition type, and we were not considering the MBR disk case, which
uses different numbers. Fixes LP:#1978127.