boot: release the new PCR handles when sealing for factory reset
It is possible that a factory reset attempt is interrupted by a reboot after
sealing the keys, which means that the PCR handles may already have been used
and have TPM resources allocated to them. In such case, we should attempt to
release the handles before sealing for factory reset again.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>