boot,device: extract SealedKey helpers from boot to device
The DA lockout reset code will need to be able to read the
`SealedKeysMethod` so this commit moves the code to read/write
the sealing methods from `boot` to `gadget/device` (just like
we did for e.g. `{Has,Read}EncryptionMarkers`).

This is split out of PR #11935 and build on top of PR #11946