bootloader/assets: add ttyS0,115200n8 to grub.cfg
This commit will help the boot speed on most PC devices and
requested by the Taipei team.

The changes to the auto-generated parts are done via:
```
$ cd boot/assets
$ go build  -o ./genasset/genasset  ./genasset/
$ ./genasset/genasset -in ./data/grub.cfg -out grub_cfg_asset.go -name grub.cfg
```
and the `Snapd-Boot-Config-Edition` for `grub.cfg` was bumped.

Note that no changes for the recovery grub configuration are done
in this commit. I'm happy to do a follwup for this.