tests: fix `core20-new-snapd-does-not-break-old-initrd` test
The nested `core20-new-snapd-does-not-break-old-initrd` installs
an old snapd and then refreshes to the new snapd and assume this
only causes a snapd restart. This assumption is incorrect, an
update like this can trigger a reboot if managed assets like
`grub.cfg` or the static commandline change. In this case a
full system reboot is performed.

This commit changes the code to deal with that by simply waiting
for the `snap install new-snapd.snap` to complete via `snap watch`
even if it take a reboot.