tests: improve `nested_get_image_name` to avoid accidential reuse
The code in `nested.sh` will reuse images with the same name. This
can lead to subtle bugs like when a nested test uses a different
model but does not specify an `NESTED_IMAGE_ID`. Because the
code in `nested_get_image_name` does not take the model into
account an image build before can get reused. This happend in
the `gadget-connections` test where depending on the test order
one test would get the wrong image.

Instead of pushing the burden to the individual tests to remember
to set `NESTED_IMAGE_ID` this commit makes `nested_get_image_name`
smarter so that it takes the model and (as a drive-by) the hash
of the custom snaps into account when building the image name.

This should prevent bugs like the one we saw with the
`gadget-connections` test.